requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
qdrant-client>=1.11.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
groq==0.9.0
python-dotenv==1.0.1
pymupdf==1.24.5
pdfplumber==0.11.0
mcp>=1.0.0
starlette==0.37.2
pydantic>=2.7.2
httpx==0.27.0
python-multipart==0.0.9